journal/requirements.txt
streamlit
boto3
Pillow
pyperclip
pandas
